Adds from_xy_coordinates_unchecked for console and curves
