+++ circuit/types/group/src/helpers/from_xy_coordinates.rs
@@ -18,10 +18,17 @@ use super::*;
 
 impl<E: Environment> Group<E> {
     /// Initializes an affine group element from a given x- and y-coordinate field element.
-    /// For safety, the resulting point is always enforced to be on the curve with constraints.
+    /// For safety, the resulting point is always enforced to be on the curve and in the subgroup.
     pub fn from_xy_coordinates(x: Field<E>, y: Field<E>) -> Self {
         // Recover point from the `(x, y)` coordinates as a witness.
         witness!(|x, y| console::Group::from_xy_coordinates(x, y))
+    }
+
+    /// Initializes an affine group element from a given x- and y-coordinate field element.
+    /// Note: The resulting point is **not** enforced to be on the curve or in the subgroup.
+    #[doc(hidden)]
+    pub fn from_xy_coordinates_unchecked(x: Field<E>, y: Field<E>) -> Self {
+        Self { x, y }
     }
 }
 
@@ -57,6 +64,31 @@ mod tests {
         }
     }
 
+    fn check_from_xy_coordinates_unchecked(
+        mode: Mode,
+        num_constants: u64,
+        num_public: u64,
+        num_private: u64,
+        num_constraints: u64,
+    ) {
+        let mut rng = TestRng::default();
+
+        for i in 0..ITERATIONS {
+            // Sample a random element.
+            let point: console::Group<<Circuit as Environment>::Network> = Uniform::rand(&mut rng);
+
+            // Inject the x- and y-coordinates.
+            let x_coordinate = Field::new(mode, point.to_x_coordinate());
+            let y_coordinate = Field::new(mode, point.to_y_coordinate());
+
+            Circuit::scope(format!("{mode} {i}"), || {
+                let affine = Group::<Circuit>::from_xy_coordinates_unchecked(x_coordinate, y_coordinate);
+                assert_eq!(point, affine.eject_value());
+                assert_scope!(num_constants, num_public, num_private, num_constraints);
+            });
+        }
+    }
+
     #[test]
     fn test_from_xy_coordinates_constant() {
         check_from_xy_coordinates(Mode::Constant, 10, 0, 0, 0);
@@ -70,5 +102,20 @@ mod tests {
     #[test]
     fn test_from_xy_coordinates_private() {
         check_from_xy_coordinates(Mode::Private, 4, 0, 14, 13);
+    }
+
+    #[test]
+    fn test_from_xy_coordinates_unchecked_constant() {
+        check_from_xy_coordinates_unchecked(Mode::Constant, 0, 0, 0, 0);
+    }
+
+    #[test]
+    fn test_from_xy_coordinates_unchecked_public() {
+        check_from_xy_coordinates_unchecked(Mode::Public, 0, 0, 0, 0);
+    }
+
+    #[test]
+    fn test_from_xy_coordinates_unchecked_private() {
+        check_from_xy_coordinates_unchecked(Mode::Private, 0, 0, 0, 0);
     }
 }
