+++ circuit/types/group/src/helpers/mod.rs
@@ -17,7 +17,6 @@
 use super::*;
 
 pub mod from_bits;
-pub mod from_unchecked;
 pub mod from_x_coordinate;
 pub mod from_xy_coordinates;
 pub mod mul_by_cofactor;
