+++ console/types/group/src/from_xy_coordinates.rs
@@ -19,6 +19,15 @@ use super::*;
 impl<E: Environment> Group<E> {
     /// Initializes a new group from the `(x, y)` affine coordinates.
     pub fn from_xy_coordinates(x: Field<E>, y: Field<E>) -> Self {
-        Self { group: E::Affine::from_coordinates((*x, *y)).into() }
+        match E::Affine::from_coordinates((*x, *y)) {
+            Some(point) => Self { group: point.into() },
+            None => E::halt("Attempted to recover an invalid group element from (x, y) coordinates"),
+        }
+    }
+
+    /// Initializes a new group from the `(x, y)` affine coordinates.
+    /// Note: The resulting point is **not** enforced to be on the curve or in the subgroup.
+    pub fn from_xy_coordinates_unchecked(x: Field<E>, y: Field<E>) -> Self {
+        Self { group: E::Affine::from_coordinates_unchecked((*x, *y)).into() }
     }
 }
